Move server related code to a separate package
We need to clearly separate layers like as follows:
cmd -> server -> client -> endpoints
